requirements.txt
django==5.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
stripe==7.0.0
pillow==10.2.0
gunicorn==21.2.0
whitenoise==6.6.0
djoser==2.2.0
